BACKEND/Project/requirements.txt
asgiref
Django
django-cors-headers
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv